add power manager and improve batteryoptim checker
