+++ app/src/main/java/com/lagradost/cloudstream3/ui/settings/SettingsFragment.kt
@@ -19,6 +19,7 @@ import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.PreferenceManager
 import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.appbar.MaterialToolbar
+import com.lagradost.cloudstream3.AcraApplication.Companion.context
 import com.lagradost.cloudstream3.R
 import com.lagradost.cloudstream3.databinding.MainSettingsBinding
 import com.lagradost.cloudstream3.mvvm.logError
@@ -153,6 +154,11 @@ class SettingsFragment : Fragment() {
 
         fun Context.isEmulatorSettings(): Boolean {
             return getLayoutInt() == 2
+        }
+
+        // phone exclusive
+        fun isTruePhone(): Boolean {
+            return !isTrueTvSettings() && !isTvSettings() && context?.isEmulatorSettings() != true
         }
 
         private fun Context.isAutoTv(): Boolean {
