+++ app/src/main/res/values/strings.xml
@@ -650,8 +650,8 @@
     <string name="yes">Yes</string>
     <string name="no">No</string>
     <string name="ok">OK</string>
-    <string name="battery_dialog_title">Battery Optimization</string>
-    <string name="battery_dialog_message">To ensure proper Downloads and Notifications, please set Battery Optimization for CloudStream app to Unrestricted.</string>
+    <string name="battery_dialog_title"> ◉ Disable Battery optimization</string>
+    <string name="battery_dialog_message">\nTo ensure proper Downloads and Notifications, please set battery optimization for CloudStream app to UNRESTRICTED.</string>
     <string name="update_notification_downloading">Downloading app update…</string>
     <string name="update_notification_installing">Installing app update…</string>
     <string name="update_notification_failed">Could not install the new version of the app</string>
